Moved files around and finished stubs for DBController
