--- src/sample/DBController.java
@@ -1,13 +1,0 @@
-public class DBController {
-
-    public static DBController myDBC;
-
-    public void insertHistory(String input, String output) {
-        // TODO implement
-    }
-
-    public ArrayList<String> getRecentHistory(int quantity) {
-        // TODO implement
-        return null;
-    }
-}
